+++ packages/core/admin/admin/src/hooks/useFetchMarketplacePlugins/index.js
@@ -1,9 +1,24 @@
 import { useQuery } from 'react-query';
-import { useNotification } from '@strapi/helper-plugin';
-import { fetchMarketplacePlugins } from './utils/api';
+import { useFetchClient, useNotification } from '@strapi/helper-plugin';
+import qs from 'qs';
+
+const MARKETPLACE_API_URL = 'https://market-api.strapi.io';
 
 const useFetchMarketplacePlugins = (notifyLoad, params) => {
   const toggleNotification = useNotification();
+  const { get } = useFetchClient();
+
+  const fetchMarketplacePlugins = async (params = {}) => {
+    const { data } = await get(`${MARKETPLACE_API_URL}/plugins`, {
+      params,
+      paramsSerializer: {
+        encode: qs.parse,
+        serialize: qs.stringify,
+      },
+    });
+
+    return data;
+  };
 
   return useQuery(['list-marketplace-plugins', params], () => fetchMarketplacePlugins(params), {
     onSuccess() {
